m8_12_01/pyproject.toml
[tool.poetry]
name = "m8-12-01"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m8_12_01"}]

[tool.poetry.dependencies]
python = "^3.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.92.0"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
sqlalchemy = "^2.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.10.5"}
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
